Fix various compiler warnings
git-svn-id: svn://svn.savannah.gnu.org/mtools/trunk@547 8fbfe7c3-7cdf-472d-ae67-d119146b2a64

+++ buffer.h
@@ -21,8 +21,8 @@
 #include "stream.h"
 
 Stream_t *buf_init(Stream_t *Next, 
-		   int size, 
-		   int cylinderSize,
-		   int sectorSize);
+		   size_t size, 
+		   size_t cylinderSize,
+		   size_t sectorSize);
 
 #endif
